Rename details.js to .jsx
